# Add AVX acceleration