@perserk 2023-03-30
@sw Thanks for the link and the patch. I changed the code and it got a bit faster, at least not slower.