@ggerganov 2023-03-30
@sw

I think this is very possible. I think this can help explain the observations in #603 
The deprecated `mad` routines were apparently utilizing much more efficiently the CPU / Memory on x86. I guess as efficient as optimized AVX2-level. But then, I removed them and effectively replaced them with a non-optimal AVX2 dot product implementation, which is probably at the level of an optimized AVX dot product.

In short, I believe the AVX2 code has room for improvements.
My suspicion is that `QK == 64` would have been more suitable for AVX2, but since Apple Silicon is highest priority, I chose `QK == 32`.