@slaren 2023-03-30
I tried adapting the same technique used here to AVX2 ([src](https://gist.github.com/slaren/869800a9e80717d0b5b9b9c6cbf5e5f0)) and it performs a little better:
```
Run on (16 X 3600 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 256 KiB (x8)
  L3 Unified 16384 KiB (x1)
Load Average: 1.11, 3.59, 2.64
------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations
------------------------------------------------------------------------
BM_ggml_vec_dot_q4_0_avx             679 ns          679 ns      1032721
BM_ggml_vec_dot_q4_0_avx2            590 ns          590 ns      1185302
BM_ggml_vec_dot_q4_0_avx2_new        525 ns          525 ns      1329767
```

While it is not exactly twice as fast, there is still a significant performance advantage in using AVX2.