@ggerganov 2023-03-31
@slaren 

If you observe good results with this and #642 , go ahead and merge both PR. I won't have much time to look today.

Btw, here is also something worth looking into: https://github.com/ggerganov/llama.cpp/discussions/205#discussioncomment-5393420